+++ src/fn/fn.ts
@@ -1,0 +1,3 @@
+import { createConditionPlugin } from '../createConditionPlugin'
+
+export const fn = createConditionPlugin(app => typeof app === 'function')