+++ src/node/node.ts
@@ -1,0 +1,3 @@
+import { createConditionPlugin } from '../createConditionPlugin'
+
+export const node = createConditionPlugin(app => app instanceof Node)