+++ src/promise/promise.ts
@@ -1,0 +1,3 @@
+import { createConditionPlugin } from '../createConditionPlugin'
+
+export const promise = createConditionPlugin(app => app instanceof Promise)