+++ src/string/string.ts
@@ -1,0 +1,3 @@
+import { createConditionPlugin } from '../createConditionPlugin'
+
+export const string = createConditionPlugin(app => typeof app === 'string')