+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1
+
+### v1.1.0 [![23.03.2022](https://img.shields.io/date/1648055838)](https://github.com/d8corp/watch-state/tree/v1.1.0)
+- added `string`, `fn`, `node` and `promise`
+- changed `async` plugin
+
 ## v1.0
 
 ### v1.0.0 [![22.02.2022](https://img.shields.io/date/1645518853)](https://github.com/d8corp/watch-state/tree/v1.0.0)